+++ app/src/main/java/com/brandonjamesyoung/levelup/fragments/QuestList.kt
@@ -4,6 +4,8 @@ import android.animation.ObjectAnimator
 import android.os.Bundle
 import android.util.Log
 import android.view.View
+
+import android.view.ViewGroup
 import android.widget.LinearLayout
 import android.widget.ProgressBar
 import android.widget.TextView
@@ -24,6 +26,7 @@ import com.brandonjamesyoung.levelup.shared.*
 import com.brandonjamesyoung.levelup.shared.LevelUpHelper.Companion.getExpToLvlUp
 import com.brandonjamesyoung.levelup.shared.ButtonHelper.Companion.convertButton
 import com.brandonjamesyoung.levelup.viewmodels.QuestListViewModel
+import com.brandonjamesyoung.levelup.views.RpgSnackbar
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
@@ -368,6 +371,18 @@ class QuestList : Fragment(R.layout.quest_list) {
         viewModel.settings.observe(viewLifecycleOwner) { settings ->
             updatePointsAcronym(settings)
         }
+
+        viewModel.message.observe(viewLifecycleOwner) { message ->
+            message.getContentIfNotHandled()?.let {
+                showSnackbar(it)
+            }
+        }
+    }
+
+    private fun showSnackbar(message: String) {
+        val view = requireView()
+        val snackbar = RpgSnackbar.make(view.rootView as ViewGroup, message)
+        snackbar.show()
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {