+++ Changelog.md
@@ -12,6 +12,12 @@ This document tracks all notables changes to the Level Up app.
 - New Quest now stores the selected icon on save
 - Quest List now shows each quest's icon or the default question mark icon if there is no icon
 
+### Fixed
+
+- Fix player input not being preserved properly when navigating from New Quest to Icon Select and back
+- Fix quest creation date being updated when editing a quest
+- Fix bug where editing a quest and selecting an icon
+
 ## 0.6.1
 
 ### Added